+++ matplotcheck/raster.py
@@ -1,9 +1,8 @@
 import numpy as np
+from .vector import VectorTester
 
-from .base import PlotTester
 
-
-class RasterTester(PlotTester):
+class RasterTester(VectorTester):
     """A PlotTester for spatial raster plots.
 
     Parameters
@@ -94,9 +93,9 @@ class RasterTester(PlotTester):
         all_label_options: list of lists
             Each internal list represents a class and said list is a list of
             strings where at least one string is expected to be in the legend
-            label for this category. Internal lists must be in the same order as
-            bins in im_expected, e.g. first internal list has the expected label
-            options for class 0.
+            label for this category. Internal lists must be in the same order
+            as bins in im_expected, e.g. first internal list has the expected
+            label options for class 0.
 
         Returns
         ----------
@@ -106,13 +105,13 @@ class RasterTester(PlotTester):
         Notes
         ----------
         First compares all_label_options against the legend labels to find
-        which element of all_label_options matches that entry. E.g. if the first
-        legend entry has a match in the first list in all_label_options, then
-        that legend entry corresponds to the first class (value 0).
-        Then the plot image array is copied and the values are set to the legend
-        label that match the values (i.e. the element in all_label_options).
-        The same is done for the expected image array. Finally those two arrays
-        of strings are compared. Passes if they match.
+        which element of all_label_options matches that entry. E.g. if the
+        first legend entry has a match in the first list in all_label_options,
+        then that legend entry corresponds to the first class (value 0).
+        Then the plot image array is copied and the values are set to the
+        legend label that match the values (i.e. the element in
+        all_label_options). The same is done for the expected image array.
+        Finally those two arrays of strings are compared. Passes if they match.
         """
         # Retrieve image array
         im_data = []
@@ -161,7 +160,43 @@ class RasterTester(PlotTester):
             im_data_labels, im_expected_labels
         ), "Incorrect legend to data relation"
 
-        ### IMAGE TESTS/HELPER FUNCTIONS ###
+        # IMAGE TESTS/HELPER FUNCTIONS
+
+    def get_plot_image(self):
+        """Returns images stored on the Axes object as a list of numpy arrays.
+
+        Returns
+        -------
+        im_data: List
+            Numpy array of images stored on Axes object.
+        """
+        im_data = []
+        if self.ax.get_images():
+            im_data = self.ax.get_images()[0].get_array()
+        assert list(im_data), "No Image Displayed"
+
+        # If image array has 3 dims (e.g. rgb image), remove alpha channel
+        if len(im_data.shape) == 3:
+            im_data = im_data[:, :, :3]
+        return im_data
+
+    def get_plot_image(self):
+        """Returns images stored on the Axes object as a list of numpy arrays.
+
+        Returns
+        -------
+        im_data: List
+            Numpy array of images stored on Axes object.
+        """
+        im_data = []
+        if self.ax.get_images():
+            im_data = self.ax.get_images()[0].get_array()
+        assert list(im_data), "No Image Displayed"
+
+        # If image array has 3 dims (e.g. rgb image), remove alpha channel
+        if len(im_data.shape) == 3:
+            im_data = im_data[:, :, :3]
+        return im_data
 
     def assert_image(
         self, im_expected, im_classified=False, m="Incorrect Image Displayed"
@@ -170,12 +205,14 @@ class RasterTester(PlotTester):
 
         Parameters
         ----------
-        im_expected: array containing the expected image data
-        im_classified: boolean, set to True image has been classified.
-            Since classified images values can be reversed or shifted and still
-            produce the same image, setting this to True will allow those
-            changes.
-        m: string error message if assertion is not met
+        im_expected: Numpy Array
+            Array containing the expected image data.
+        im_classified: boolean
+            Set to True image has been classified. Since classified images
+            values can be reversed or shifted and still produce the same image,
+            setting this to True will allow those changes.
+        m: string
+            String error message if assertion is not met.
 
         Returns
         ----------