+++ CHANGELOG.md
@@ -5,9 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
+* Removed the `m2r` package from MatPlotCheck (@nkorinek, #247)
+* Made `assert_string_contains()` accept a string instead of a list (@nkorinek, #53)
+* Added functions to get and assert the midpoint values of bins in a histogram (@nkorinek, #184)
+* Created tests for the autograde module (@nkorinek, #105)
+* Created tests for the vector module (@nkorinek, #209)
+* Created functions to test point geometries in VectorTester (@nkorinek, #176)
+* made `assert_string_contains()` accept correct strings with spaces in them (@nkorinek, #182)
+* added contributors file and updated README to remove that information (@nkorinek, #121)
+
+## [0.1.2]
+* Adding flake8 for format and other syntax issues! yay! (@lwasser, #195)
 * Created a vignette covering the testing of histograms (@ryla5068, #149)
+* Created `get_plot_image()` function for the RasterTester object (@nkorinek, #192)
+* Allowed `assert_polygons()` to accept GeoDataFrames and added tests (@nkorinek, #175)
+* raster test inherits from vector to allow for multi layer plot testing (@lwasser, #75)
 
 ## [0.1.1]
 * Added test for bin heights of histograms (@ryla5068, #124)